Support older ffmpeg libraries.
Ubuntu 16.04 only comes with ffmpeg 2.8 which has a different API than
version 3.4.  In particular, an AVStream doesn't have
AVCodecParameters* codecpar but an AVCodecContext* codec instead.
Also, the bit stream filter API is completely different.

This commit detects these differences in the cmake files and changes
the code accordingly. It seems to now work with both versions of
ffmpeg.

Fixes #4