Make sure we have a vid_decoder_ before sending it requests.
Doing so means we can't send the vid_decoder_ a request until we are
in the read_file loop.  This means in the decoder we can't wait around
for a request to see if the frame we are getting is one we want, since
that would stall the loop that sends it a request.  So in the case, we
ditch those frames. In our normal use case it is highly unlikely those
are the frame we want anyway.