Fix compiling with -O2 and cleanup code dealing with types.
Previously we were relying on C functions calling into template
functions to instantiate the template functions. This doesn't work
with -O2 since the template functions are just inlined, not
instantiated. So we need to explicitely instantiate those functions.

Instead of doing it manually, I pulled in boost::mp11 and did some
metaprogramming.  First, we define a mp11 map from C enums of
supported types to actual types. This gives us a list of types we want
to support.  For each of those types we instantiate a struct that
holds pointers to the functions we want instantiated.

Having that map from C enum to C++ type also allows us to define the
variant types from that map, as well as use them in the C wrapper
functions. So adding new types later will simply be a matter of adding
an entry into the C enum, then adding an entry into the mp11 map to
indicate which real type to use.

Since we use don't really use the meta information in a sequence very
much, pare down the list of types supported in metadata for the
sequence to just int and string.